requirements.txt
numpy
python-socketio
aiohttp
requests
boto3
pyyaml
datasets
scipy
transformers
sklearn
torch
torchvision